VP-2527: Change cache expiration for blob changes (#48)
* VP-2527: Change cache expiration for blob changes

- Added changesSince date to the cache key;
- Removed SetAbsoluteExpiration(TimeSpan.FromMinutes(5)) as it could clear cache many times even while one big task execution in progress;
- Added BlobChanges caching based on path and changesSince date, expiration is done on task processing completion;

* VP-2527: Set 'LastRun' time to the time before processing

This should be done to prevent "changes loss", if they were done in the interval between getting changes from the blob and all tasks completion save, when the date was updated before.

* VP-2527: Execute generation task and save results one by one

It is done to allow completed tasks log their results and not to wait for other long running tasks (e.g. when call ProcessAll). Thus even if all tasks job is still in process, or failed, the already generated images will not be generated second time (e.g. if run specific tasks manually or restart ProcessAll tasks) - thanks to updated LastRun date.

* VP-2527: Fix possible invalid cast exception

* VP-2527: Add tests for images changes caching

* VP-2527: Add tests on caching for ThumbnailGenerationProcessor

* VP-2527: Fix typo

* VP-2527: Fix review comment by removing redundant hashset